Fix stream and commons util static analysis issues

+++ core/commons/src/main/java/com/bezirk/middleware/core/util/Hashids.java
@@ -372,7 +372,7 @@ public class Hashids {
 
         for (int i = 0; i < input.length(); i++) {
             pos = alphabet.indexOf(input_arr[i]);
-            number += pos * (long) Math.pow(alphabet.length(), input.length() - i - 1);
+            number += pos * (long) Math.pow(alphabet.length(), (double) input.length() - (double) i - 1);
         }
 
         return number;
